+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file. Based on [K
 
 ## [Unreleased]
 
+### Added
+- Historic results in Ranking page
+
 ## [0.2.0] - 2021-07-27
 
 ### Added